requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
google-generativeai
python-dotenv
